+++ templates/load-more-button.php
@@ -4,5 +4,5 @@ if (!defined('ABSPATH')) {
 }
 //load more button ?>
 <div class="quick-ajax-load-more">
-    <button type="button" class="quick-ajax-load-more-button quick-ajax-button" data-button="quick-ajax-load-more">QUICK_AJAX_LABEL</button>
+    <button type="button" class="quick-ajax-load-more-button quick-ajax-button" data-button="quick-ajax-load-more">[[QAPL::LOAD_MORE_LABEL]]</button>
 </div>