+++ templates/post-items/post-item.php
@@ -16,12 +16,13 @@ if (!defined('ABSPATH')) {
             echo wp_get_attachment_image($thumbnail_id, 'full');
             echo '</div>';
         }
-        ?>    
+        ?>
         <div class="post-title <?php echo has_post_thumbnail() ? '' : 'no-image'; ?>">
             <h3><?php echo esc_html(get_the_title()); ?></h3>
         </div>
         <div class="post-desc">
             <p><?php echo esc_html(wp_trim_excerpt()); ?></p>
+            <p>[[QAPL::READ_MORE_LABEL]]</p>
         </div>
     </a>
 </div>